feat : 보너스 번호 추가된 게임 진행 구현
