+++ README.md
@@ -126,6 +126,6 @@ System.out.println(result); // print 6
   - [X] 당첨번호 5개와 보너스 번호가 일치하면 2등이 된다
 - [X] 당첨 통계에 2등이 추가된다
   - [X] 당첨 통계에 2등 추가된다
-- [ ] 게임진행
-  - [ ] 보너스 볼을 입력 받는다
-  - [ ] 당첨 통계에 2등이 추가되어 출력된다
+- [X] 게임진행
+  - [X] 보너스 볼을 입력 받는다
+  - [X] 당첨 통계에 2등이 추가되어 출력된다
